requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
PyJWT
passlib[bcrypt]
python-multipart
fpdf2